Add cpp config. Move other Map types to MapTypes.hpp. Clean up includes
