+++ src/Shared/Maps/Map.hpp
@@ -3,97 +3,14 @@
 
 #include <string>
 #include <vector>
-#include "Shared/Entities/EntityManager.hpp"
-#include "Shared/Util/ResourcePool.hpp"
+
 #include "Shared/Util/Vector2d.hpp"
-#include "Shared/Scripts/ScriptManager.hpp"
-#include "Shared/Scripts/Specialized/MapScript.hpp"
 #include "Shared/Maps/Tileset.hpp"
 #include "Shared/Util/File.hpp"
 #include "Shared/Maps/Weather.hpp"
-#include "Shared/Entities/Instances/ItemEntity.hpp"
+#include "Shared/Maps/MapTypes.hpp"
 
 class Playlist;
-class Map;
-
-/**
- * Structure to store a tile in the map
- *
- * \ingroup World
- */
-struct Tile {
-    /**
-     * Creates an empty tile
-     */
-    Tile() {
-        nonZero = false;
-        isAnim = false;
-        delA = false;
-    }
-    bool isAnim, nonZero, delA;
-    Animation* anim;
-    sf::Sprite spr;
-    int id;
-};
-
-/**
- * Structure to store a light in the map
- *
- * \ingroup World
- */
-struct Light {
-    sf::Vector2f position;
-    int radius, threshold;
-};
-
-/**
- * Structure to store a script in the map
- *
- * \ingroup World
- */
-struct MapEvent {
-    Script::Ptr script;
-    std::string scriptStr;
-    sf::Vector2i position;
-    sf::Vector2i size;
-    int maxRuns, trigger; //0 = on load, 1 = step in range, 2 = step out of range, 3 = step in or out of range, 4 = in range
-    int runs;
-};
-
-/**
- * Struct to store entity spawn data
- */
-struct EntitySpawn {
-	std::string name;
-	EntityPosition position;
-};
-
-/**
- * Structure to store details of a spawner that is in a map
- *
- * \ingroup Map
- */
-struct MapSpawner {
-	sf::Vector2i position;
-	int coolDown, frequency;
-	int numActiveLimit, spawnLimit;
-	std::string templateFile;
-	//TODO - add actual spawner data or make a separate class
-};
-
-/**
- * Structure to store details of items in the map
- *
- * \ingroup Map
- */
-struct MapItem {
-	int itemId, mapId;
-	sf::Vector2i position;
-
-private:
-	Entity::Ptr ie;
-	friend class Map;
-};
 
 class Map {
 	std::string name, uniqueName;
@@ -176,11 +93,7 @@ public:
      *
      * \param file The name of the map to load. Use "LastMap" to load the previous map
      * \param tileset A reference to the Tileset object
-     * \param entityManager A pointer to the EntityManager
-     * \param se A pointer to the SoundEngine object
      * \param player A pointer to the player Entity object
-     * \param spName The name of the spawn to put the player at. Use "prev" to put the player where they were in the previous map
-     * \param plst A pointer to the Playlist object to update, if any
      */
     Map(std::string file, Tileset& tileset, Entity::Ptr player = nullptr);
